+++ src/gws_biota/compound/compound_position.py
@@ -3,7 +3,7 @@
 # The use and distribution of this software is prohibited without the prior consent of Gencovery SAS.
 # About us: https://gencovery.com
 
-from peewee import CharField, DecimalField
+from peewee import CharField, FloatField
 from gws_core.model.typing_register_decorator import typing_registrator
 from gws_core import BadRequestException
 from ..base.base import Base
@@ -13,9 +13,9 @@ from .compound_position_data import COMPOUND_POSITION_DATA
 @typing_registrator(unique_name="CompoundPosition", object_type="MODEL", hide=True)
 class CompoundPosition(Base):
     chebi_id = CharField(null=True, index=True)
-    x = DecimalField(null=True, index=True)
-    y = DecimalField(null=True, index=True)
-    z = DecimalField(null=True, index=True)
+    x = FloatField(null=True, index=True)
+    y = FloatField(null=True, index=True)
+    z = FloatField(null=True, index=True)
 
     @classmethod
     def get_by_chebi_id(cls, chebi_id: str):
